--- a/optimum/habana/transformers/generation/utils.py
+++ b/optimum/habana/transformers/generation/utils.py
@@ -597,7 +597,8 @@ class GaudiGenerationMixin(GenerationMixin):
                 "llama",
                 "mistral",
                 "falcon",
-            ], "reuse_cache only supported by llama, mistral and falcon at the moment"
+                "gptj",
+            ], "reuse_cache only supported by llama, mistral, falcon and gptj at the moment"
             if not generation_config.bucket_internal:
                 assert (
                     generation_config.bucket_size <= 0