--- a/optimum/habana/transformers/models/gptj/__init__.py
+++ b/optimum/habana/transformers/models/gptj/__init__.py
@@ -1,6 +1,6 @@
 from .modeling_gptj import (
     GaudiGPTJAttention,
     GaudiGPTJBlock,
-    GaudiGPTJModel,
     GaudiGPTJForCausalLM,
+    GaudiGPTJModel,
 )