# What does this PR do?

Add GPT-J FP8 supported

# Command Lines:
- Measure the tensor quantization statistic on `EleutherAI/gpt-j-6b`
  > ```
  > QUANT_CONFIG=./quantization_config/maxabs_measure.json python run_generation.py \
  >   --model_name_or_path EleutherAI/gpt-j-6b \
  >   --use_hpu_graphs \
  >   --use_kv_cache \
  >   --limit_hpu_graphs \
  >   --max_input_tokens 128 \
  >   --max_new_tokens 128 \
  >   --batch_size 1 \
  >   --bf16
  > ```
- Quantize the model based on previous measurements for `EleutherAI/gpt-j-6b`

  > ```
  > QUANT_CONFIG=./quantization_config/maxabs_quant.json python run_generation.py \
  > --model_name_or_path EleutherAI/gpt-j-6b \
  > --use_hpu_graphs \
  > --use_kv_cache \
  > --limit_hpu_graphs \
  > --max_input_tokens 128 \
  > --max_new_tokens 128 \
  > --batch_size 256 \
  > --bf16 \
  > --fp8
  > ```