@kplau1128 2024-04-24
> @kplau1128 , Can you please add some command lines to the PR description, that was used to test this

@ssarkar2 Command lines for measure and quantization added on description.