@ssarkar2 2024-04-25
@kplau1128 , This PR seems to be in "Draft" mode. Can you please more it to regular PR

![image](https://github.com/huggingface/optimum-habana/assets/15640036/b996c07c-bd41-4382-8e7c-0ae692a27c73)