@kplau1128 2024-05-01
> Can you please run bf16 tests and make sure you are getting similar numbers with and without this PR as well. for example a couple of configs like 128->128, 128->2048, 2048->128, 2048->2048 (i->o, meaning max_input_tokens=i and max_new_tokens=o)
> 
> also can you please run `make style` at the topmost folder, because the PR is failing style check

Style checked.

Also add commit with added `--reuse_cache` option support.