@kplau1128 2024-05-06
Test with this PR seeing throughput improve, but still low.

Also see main  GPT-J code already has `The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3` issue, when either `max_input_tokens` or `max_new_tokens` set to `2048` 

### Without this PR:
|Model|batch_size|max_input_tokens|max_new_tokens|Throughput (tokens/second)|Number of HPU graphs|Memory allocated (GB)|Max memory allocated (GB)|Total memory available (GB)|Graph compilation duration (seconds)|Error|
|---|---|---|---|---|---|---|---|---|---|---|
|GPT-J-6B Measure|1|128|128|147.26|11|11.52|11.68|94.62|8.71||
|GPT-J-6B Quantization|1024|128|128|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::2147483648 (2048)MB|
|GPT-J-6B Quantization|640|128|128|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::33030144000 (31500)MB|
|GPT-J-6B Quantization|320|128|128|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::16515072000 (15750)MB|
|GPT-J-6B Quantization|256|128|128|4978.55|15|40.21|86.68|94.62|80.36||
|GPT-J-6B Quantization|120|2048|128|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|120|1024|128|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::27869184000 (26578.1)MB|
|GPT-J-6B Quantization|120|368|128|2823.54|15|37.01|76.51|94.62|78.28||
|GPT-J-6B Quantization|120|128|2048|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|120|128|1024|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::27869184000 (26578.1)MB|
|GPT-J-6B Quantization|120|128|624|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::18192384000 (17349.6)MB|
|GPT-J-6B Quantization|120|128|512|3706.45|15|46.02|94.62|94.62|113.97||
|GPT-J-6B Quantization|120|128|256|4687.05|15|30.01|62.27|94.62|77.36||
|GPT-J-6B Quantization|64|2048|128|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|64|1024|128|1183.81|15|44.39|91.28|94.62|103.23||
|GPT-J-6B Quantization|64|512|128|2032.33|15|27.35|54.08|94.62|70.22||
|GPT-J-6B Quantization|64|2048|2048|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (4096) at non-singleton dimension 3|
|GPT-J-6B Quantization|64|1024|1024|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::26424115200 (25200)MB|
|GPT-J-6B Quantization|64|512|512|2514.55|15|40.12|82.57|94.62|113.1||


### With this PR
|Model|batch_size|max_input_tokens|max_new_tokens|Throughput (tokens/second)|Number of HPU graphs|Memory allocated (GB)|Max memory allocated (GB)|Total memory available (GB)|Graph compilation duration (seconds)|Error|
|---|---|---|---|---|---|---|---|---|---|---|
|GPT-J-6B Measure|1|128|128|87.77|21|11.52|11.57|94.62|6.44||
|GPT-J-6B Quantization|1024|128|128|||||||RuntimeError: [Rank:0] FATAL ERROR :: MODULE:PT_DEVMEM Allocation failed for size::52848230400 (50400)MB|
|GPT-J-6B Quantization|640|128|128|5822.32|74|56.74|94.62|94.62|115.87||
|GPT-J-6B Quantization|320|128|128|5654.14|74|31.34|65.78|94.62|85.91||
|GPT-J-6B Quantization|256|128|128|5717.59|74|26.27|53.58|94.62|80.37||
|GPT-J-6B Quantization|120|2048|128|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|120|1024|128|1533.06|74|48.53|78.33|94.62|103.88||
|GPT-J-6B Quantization|120|368|128|3323.75|74|24.32|48.44|94.62|86.0||
|GPT-J-6B Quantization|120|128|2048|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|120|128|1024|3910.81|74|48.53|78.33|94.62|169.21||
|GPT-J-6B Quantization|120|128|624|4892.65|74|33.77|70.57|94.62|118.27||
|GPT-J-6B Quantization|120|128|512|5022.39|74|29.64|61.05|94.62|101.95||
|GPT-J-6B Quantization|120|128|256|5714.56|74|20.19|38.94|94.62|80.3||
|GPT-J-6B Quantization|64|2048|128|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (2176) at non-singleton dimension 3|
|GPT-J-6B Quantization|64|1024|128|1464.83|74|28.66|58.54|94.62|84.13||
|GPT-J-6B Quantization|64|512|128|2402.4|74|18.58|35.04|94.62|73.62||
|GPT-J-6B Quantization|64|2048|2048|||||||RuntimeError: The size of tensor a (2048) must match the size of tensor b (4096) at non-singleton dimension 3|
|GPT-J-6B Quantization|64|1024|1024|2431.85|74|48.18|76.33|94.62|140.18||
|GPT-J-6B Quantization|64|512|512|3365.81|74|26.15|52.71|94.62|91.72||